# api/tasks.openapi.yaml
openapi: 3.0.3
info:
  title: NSW Sri Lanka / Tasks API
  description: |
    OpenAPI 3.0 specification for the task-orchestrator HTTP surface: reading
    a single task's rendered "zone view" and advancing it by submitting a
    step. A task is one node of a consignment's workflow; access is derived
    from the caller's role-tied ownership of the task's consignment.

    All routes require a JWT bearer token (see `securitySchemes.BearerAuth`)
    carrying the scope named on each operation. Every error response uses the
    shared `ErrorResponse` envelope (`{"error": "...", "correlationId": "..."}`).
  version: 1.0.0

security:
  - BearerAuth: []

paths:
  /api/v1/tasks/{id}:
    get:
      summary: Get a task's rendered zone view
      description: |
        Returns the ZoneView payload for a single task, scoped to the
        caller: they must own the task's consignment in a role the task's
        render config admits. A task the caller may not read is reported the
        same as one that does not exist (404), so the response cannot be
        used to probe which task IDs exist.

        Each value the task's render config declares under `files` is
        replaced by a **file reference** issued to the caller: an opaque
        string to pass to `GET /api/v1/storage/{key}`, or back in a
        submission, never the stored file's key. A reference expires after
        the deployment's `storage.tokenTTLSeconds`; refetch the task for
        fresh ones.

        Requires scope `nsw:task:read`.
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskID'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneView'
        '400':
          description: Missing task ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. No usable principal on the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden. The bearer token lacks scope `nsw:task:read`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: |
            No such task, or the caller may not read it (indistinguishable
            from "not found" by design).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: |
            Internal error resolving access, issuing file references (a
            render config whose `files` does not parse included), or
            assembling the view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks/{id}/steps/{stepId}:
    post:
      summary: Advance a task by completing a step (portal)
      description: |
        Submits a command (and optional payload) to advance the task's
        underlying workflow. The request body must contain exactly one JSON
        value; trailing data after it is rejected.

        `{stepId}` is the step being completed: the `step_id` the task's zone
        view reported. Only that step is completed; a submission for a step
        the task has since left is rejected with 409 and changes nothing.
        External reviewers call back on `POST /api/v1/callbacks/{token}`
        instead.

        Each `payload` field the task's render config declares under `files`
        must hold a file reference issued to the caller, from an upload or
        the task's zone view; the stored value it stands for is what gets
        stored. For now, a value that is not a reference is stored as sent.

        Requires scope `nsw:task:write`.
      operationId: completeTaskStep
      parameters:
        - $ref: '#/components/parameters/TaskID'
        - $ref: '#/components/parameters/StepID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskStepRequest'
      responses:
        '204':
          description: |
            No Content. The workflow accepted the step; the task shows
            ADVANCING (or its next step) from then on, so refetch it rather
            than assume its state.
        '400':
          description: |
            Missing task or step ID, malformed JSON, unexpected trailing data,
            a missing/empty `command`, or a `payload` that sets the reserved
            `__command` key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: |
            The caller may not perform this action on this task, or a
            declared file field holds a file reference issued to someone
            else.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            `{stepId}` is not the task's active step: the task has moved on (or
            completed) since the caller read it. Refetch the task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: |
            A declared file field holds a file reference that has expired.
            Refetch the task, or upload the file again, and resubmit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: |
            Internal error resolving the submission's file references (a
            render config whose `files` does not parse included) or
            completing the step.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/callbacks/{token}:
    post:
      summary: Complete the step an external reviewer was dispatched for
      description: |
        The route an external reviewer (an OGA portal or agency) calls back
        on. `{token}` is the opaque `callbackToken` from the dispatch envelope
        (the reviewer posts to `{serviceUrl}/{callbackToken}`). It names one
        step of one task, so a callback that arrives after the task has moved
        on is rejected with 409 instead of completing a later step. Treat it
        as an opaque string; it is also the reviewer's idempotency key (the
        same on every retry of one dispatch, different for every step).

        The body is the same envelope as the portal's
        `POST /api/v1/tasks/{id}/steps/{stepId}`, and must contain exactly one
        JSON value. Its declared file fields follow the same rule: each must
        hold a file reference issued to the caller's client, such as one the
        dispatch carried or an upload returned.

        Requires scope `nsw:task:write`.
      operationId: completeTaskStepByCallbackToken
      parameters:
        - name: token
          in: path
          required: true
          description: The `callbackToken` from the dispatch envelope.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskStepRequest'
      responses:
        '204':
          description: No Content. The workflow accepted the step.
        '400':
          description: |
            A malformed token, malformed JSON, unexpected trailing data, a
            missing/empty `command`, or a `payload` that sets the reserved
            `__command` key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: |
            The caller may not perform this action on this task, or a
            declared file field holds a file reference issued to someone
            else.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            The token's step is no longer the task's active step: this callback
            was already applied, or the task has moved on.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: |
            A declared file field holds a file reference that has expired.
            Refetch the task, or upload the file again, and resubmit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: |
            Internal error resolving the submission's file references (a
            render config whose `files` does not parse included) or
            completing the step.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    TaskID:
      name: id
      in: path
      required: true
      schema:
        type: string
    StepID:
      name: stepId
      in: path
      required: true
      description: The step being completed (the task's `step_id`).
      schema:
        type: string

  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "task not found"
        correlationId:
          type: string
          example: "1d67eb9ce46b7b08c6bcf1bdd91de5d5"

    CompleteTaskStepRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          minLength: 1
          description: The step command to submit, e.g. "submit".
          example: "submit"
        payload:
          type: object
          description: |
            Command-specific data. Genuinely dynamic per task type; must not
            set the reserved `__command` key, which the server adds itself.
          additionalProperties: true
          not:
            required:
              - __command

    HandleClaim:
      type: object
      description: One zone's binding of a command to a renderer-defined screen element.
      required:
        - command
        - label
      properties:
        command:
          type: string
        label:
          type: string
        element:
          type: string

    EnrichedComponent:
      type: object
      description: The per-zone wire shape after merging projector output with the zone's legal handles.
      required:
        - id
        - type
        - payload
      properties:
        id:
          type: string
          description: The section key from the task's render config. Unique within a view.
        title:
          type: string
          description: Heading for the section. Omitted when the render config gives none.
        type:
          type: string
        handles:
          type: array
          description: Present (non-empty) iff this zone is interactive in the task's current state.
          items:
            $ref: '#/components/schemas/HandleClaim'
        payload:
          description: Renderer-defined payload for this zone's component type.
          additionalProperties: true

    ZoneView:
      type: object
      description: |
        The wire shape the trader-app's zone renderer consumes. `view` is the
        ordered list of EnrichedComponent to render, in the order the task's
        render config lays them out for the current state; there is no separate
        top-level actions list — actions ship inside their claiming zone's `handles`.
      required:
        - task_id
        - task_type
        - state
        - view
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
        task_type:
          type: string
        state:
          type: string
          description: e.g. PENDING_USER, QUEUED_EXTERNALLY, COMPLETED.
        step_id:
          type: string
          description: |
            The step the task is on. A submission is posted to
            `/api/v1/tasks/{id}/steps/{step_id}`. Omitted before the task's
            first step starts.
        version:
          type: integer
          format: int64
          description: Increases every time the task moves on. Read-only.
        view:
          type: array
          description: Sections in render order.
          items:
            $ref: '#/components/schemas/EnrichedComponent'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time